trading_ai_starter/requirements.txt
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.4.2
xgboost>=2.0.3
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.1
python-dotenv>=1.0.1
MetaTrader5>=5.0.45
requests>=2.32.2
